Refactored TaxNumberGenerator into its own class. Use Random.ArrayElement utility helper. Updated README extensions docs.
